requirements.txt
absl-py==1.4.0
beartype==0.13.1
cola-ml==0.0.1
gpjax==0.7.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
jax==0.4.14
jaxlib==0.4.14
jaxtyping==0.2.21
omegaconf==2.3.0
optax==0.1.7